implemented one-to-many for user-notes

+++ src/app.js
@@ -25,14 +25,6 @@ passport.deserializeUser(function(userStr, done) {
   done(null, JSON.parse(userStr));
 });
 
-function authenticate(req, res, next) {
-  if (req.user) {
-    return next();
-  } else {
-    return passport.authenticate('google')(req, res);
-  }
-}
-
 // configuring express
 
 app.configure(function() {
@@ -56,9 +48,19 @@ app.configure(function() {
   });
 });
 
-app.get(/\/notes\/(\w+)/, function (req, res) {
+function authenticate(req, res, next) {
+  if (req.user) {
+    return next();
+  } else {
+    return passport.authenticate('google')(req, res);
+  }
+}
+
+app.get(/\/notes\/(\w+)/, 
+  authenticate,
+  function (req, res) {
   var id = req.params[0];
-  db.getEntity('notes', id).then(function (data) {
+  db.getEntity('notes', { _id: db.oid(id), userId: req.user.id }).then(function (data) {
     res.send(data);
   }, function (err) {
     res.writeHead(400, "Bad request");
@@ -69,7 +71,7 @@ app.get(/\/notes\/(\w+)/, function (req, res) {
 app.get(/\/notes/, 
   authenticate,
   function (req, res) {
-  db.getEntities('notes').then(function (data) {
+  db.getEntities('notes', { userId: req.user.id }).then(function (data) {
     res.send(data);
   }, function (err) {
     res.writeHead(400, "Bad request");
@@ -81,7 +83,7 @@ app.delete(/\/notes\/(\w+)/,
   authenticate,
   function (req, res) {
   var id = req.params[0];
-  db.deleteEntity('notes', id).then(function () {
+  db.deleteEntity('notes', { _id: db.oid(id), userId: req.user.id }).then(function () {
     res.send(true);
   }, function (err) {
     res.writeHead(400, "Bad request");
@@ -92,7 +94,7 @@ app.delete(/\/notes\/(\w+)/,
 app.post(/\/notes/, 
   authenticate,
   function (req, res) {
-  db.insertEntity('notes', { text: req.query.text }).then(function (data) {
+  db.insertEntity('notes', { userId: req.user.id, text: req.query.text }).then(function (data) {
     res.send(data[0]);
   }, function (err) {
     res.writeHead(400, "Bad request");
@@ -104,7 +106,7 @@ app.put(/\/notes\/(\w+)/,
   authenticate,
   function (req, res) {
   var id = req.params[0];
-  db.putEntity('notes', id, { text: req.query.text }).then(function (data) {
+  db.putEntity('notes', { _id: db.oid(id), userId: req.user.id }, { text: req.query.text }).then(function (data) {
     res.send({ _id: id, text: req.query.text });
   }, function (err) {
     res.writeHead(400, "Bad request");
